+++ samplesheets/plugins.py
@@ -390,14 +390,14 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
         :param project: Project object to limit update to (optional)
         :param user: User object to denote user triggering the update (optional)
         """
-        irods_backend = get_backend_api('omics_irods')
-        cache_backend = get_backend_api('sodar_cache')
+        try:
+            cache_backend = get_backend_api('sodar_cache')
+            irods_backend = get_backend_api('omics_irods')
 
-        if (
-            not cache_backend
-            or not irods_backend
-            or not irods_backend.test_connection()
-        ):
+        except Exception:
+            return
+
+        if not cache_backend or not irods_backend:
             return
 
         # Study sub-app plugins